(feat) Make Form reusable
