+++ strategy-builder-gui/src/pages/Settings/composables/useSettingsForm.ts
@@ -1,14 +1,15 @@
 import { computed } from 'vue';
 
 import { $settingsForm } from '../stores/settingsForm';
+import { FormList } from '../types';
 
-export const useSettingsForm = () => {
+export const useSettingsForm = (form: FormList) => {
   const values = computed(() =>
     Object.keys($settingsForm).reduce(
-      (acc, key) => ({ ...acc, [key]: $settingsForm[key].value.value }),
+      (acc, key) => ({ ...acc, [key]: $settingsForm[form][key].value.value }),
       {},
     ),
   );
 
-  return { settingsForm: $settingsForm, values };
+  return { settingsForm: $settingsForm[form], values };
 };
